# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Mangaweb",
        "title": "Mangaweb3 API",
        "contact": {},
        "version": "3.0"
    },
    "paths": {
...
        "/view/populate_tags": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/view.populateTagsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/view.populateTagsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    }
                }
            }
        },
        "/view/set_favorite": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/view.setFavoriteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/view.setFavoriteResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    }
                }
            }
        },
        "/view/update_cover": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/view.updateCoverRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/view.updateCoverResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "ent.History": {
            "type": "object",
            "properties": {
                "create_time": {
                    "description": "CreateTime holds the value of the \"create_time\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the HistoryQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.HistoryEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.HistoryEdges": {
            "type": "object",
            "properties": {
                "item": {
                    "description": "Item holds the value of the item edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.Meta"
                        }
                    ]
                },
                "user": {
                    "description": "User holds the value of the user edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.User"
                        }
                    ]
                }
            }
        },
        "ent.Meta": {
            "type": "object",
            "properties": {
                "active": {
                    "description": "Active holds the value of the \"active\" field.",
                    "type": "boolean"
                },
                "container_type": {
                    "description": "ContainerType holds the value of the \"container_type\" field.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/meta.ContainerType"
                        }
                    ]
                },
                "create_time": {
                    "description": "CreateTime holds the value of the \"create_time\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the MetaQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.MetaEdges"
                        }
                    ]
                },
                "favorite": {
                    "description": "Favorite holds the value of the \"favorite\" field.",
                    "type": "boolean"
                },
                "file_indices": {
                    "description": "FileIndices holds the value of the \"file_indices\" field.",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "hidden": {
                    "description": "Hidden holds the value of the \"hidden\" field.",
                    "type": "boolean"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "read": {
                    "description": "Read holds the value of the \"read\" field.",
                    "type": "boolean"
                },
                "thumbnail_height": {
                    "description": "ThumbnailHeight holds the value of the \"thumbnail_height\" field.",
                    "type": "integer"
                },
                "thumbnail_index": {
                    "description": "ThumbnailIndex holds the value of the \"thumbnail_index\" field.",
                    "type": "integer"
                },
                "thumbnail_width": {
                    "description": "ThumbnailWidth holds the value of the \"thumbnail_width\" field.",
                    "type": "integer"
                },
                "thumbnail_x": {
                    "description": "ThumbnailX holds the value of the \"thumbnail_x\" field.",
                    "type": "integer"
                },
                "thumbnail_y": {
                    "description": "ThumbnailY holds the value of the \"thumbnail_y\" field.",
                    "type": "integer"
                }
            }
        },
        "ent.MetaEdges": {
            "type": "object",
            "properties": {
                "favorite_of_user": {
                    "description": "FavoriteOfUser holds the value of the favorite_of_user edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.User"
                    }
                },
                "histories": {
                    "description": "Histories holds the value of the histories edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.History"
                    }
                },
                "progress": {
                    "description": "Progress holds the value of the progress edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Progress"
                    }
                },
                "tags": {
                    "description": "Tags holds the value of the tags edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Tag"
                    }
                }
            }
        },
        "ent.Progress": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ProgressQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.ProgressEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "item_id": {
                    "description": "ItemID holds the value of the \"item_id\" field.",
                    "type": "integer"
                },
                "page": {
                    "description": "Page holds the value of the \"page\" field.",
                    "type": "integer"
                },
                "user_id": {
                    "description": "UserID holds the value of the \"user_id\" field.",
                    "type": "integer"
                }
            }
        },
        "ent.ProgressEdges": {
            "type": "object",
            "properties": {
                "item": {
                    "description": "Item holds the value of the item edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.Meta"
                        }
                    ]
                },
                "user": {
                    "description": "User holds the value of the user edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.User"
                        }
                    ]
                }
            }
        },
        "ent.Tag": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the TagQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.TagEdges"
                        }
                    ]
                },
                "favorite": {
                    "description": "Favorite holds the value of the \"favorite\" field.",
                    "type": "boolean"
                },
                "hidden": {
                    "description": "Hidden holds the value of the \"hidden\" field.",
                    "type": "boolean"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.TagEdges": {
            "type": "object",
            "properties": {
                "favorite_of_user": {
                    "description": "FavoriteOfUser holds the value of the favorite_of_user edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.User"
                    }
                },
                "meta": {
                    "description": "Meta holds the value of the meta edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Meta"
                    }
                }
            }
        },
        "ent.User": {
            "type": "object",
            "properties": {
                "active": {
                    "description": "Active holds the value of the \"active\" field.",
                    "type": "boolean"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UserQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.UserEdges"
                        }
                    ]
                },
                "email": {
                    "description": "Email holds the value of the \"email\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.UserEdges": {
            "type": "object",
            "properties": {
                "favorite_items": {
                    "description": "FavoriteItems holds the value of the favorite_items edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Meta"
                    }
                },
                "favorite_tags": {
                    "description": "FavoriteTags holds the value of the favorite_tags edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Tag"
                    }
                },
                "histories": {
                    "description": "Histories holds the value of the histories edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.History"
                    }
                },
                "progress": {
                    "description": "Progress holds the value of the progress edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Progress"
                    }
                }
            }
        },
        "errors.Error": {
            "type": "object",
            "properties": {
                "cause": {},
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "message": {
                    "type": "string",
                    "example": "unknown error."
                }
            }
        },
...
        "meta.ContainerType": {
            "type": "string",
            "enum": [
                "zip",
                "zip",
                "directory"
            ],
            "x-enum-varnames": [
                "DefaultContainerType",
                "ContainerTypeZip",
                "ContainerTypeDirectory"
            ]
        },
        "meta.CropDetails": {
            "type": "object",
            "properties": {
                "height": {
                    "type": "integer"
                },
                "width": {
                    "type": "integer"
                },
                "x": {
                    "type": "integer"
                },
                "y": {
                    "type": "integer"
                }
            }
        },
...
        "view.populateTagsRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "view.populateTagsResponse": {
            "type": "object",
            "properties": {
                "favorite": {
                    "type": "boolean"
                },
                "indices": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "name": {
                    "type": "string"
                },
                "request": {
                    "$ref": "#/definitions/view.populateTagsRequest"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Tag"
                    }
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "view.setFavoriteRequest": {
            "type": "object",
            "properties": {
                "favorite": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "view.setFavoriteResponse": {
            "type": "object",
            "properties": {
                "favorite": {
                    "type": "boolean"
                },
                "request": {
                    "$ref": "#/definitions/view.setFavoriteRequest"
                }
            }
        },
        "view.updateCoverRequest": {
            "type": "object",
            "properties": {
                "crop_details": {
                    "$ref": "#/definitions/meta.CropDetails"
                },
                "index": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "view.updateCoverResponse": {
            "type": "object",
            "properties": {
                "request": {
                    "$ref": "#/definitions/view.updateCoverRequest"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
...
}